Merge dev into main

+++ src/shared/api/auth/authService.ts
@@ -1,50 +1,51 @@
 import { type AxiosError, type AxiosResponse } from 'axios'
-
 import { $api } from 'shared/api/api'
+
 import {
-    type PasswordRecoveryModel,
-    type UserCreatePasswordModel,
-    type UserAuthModel,
-    type UserLoginModel,
-    type UserRegistrationModel,
-    type UseResendLinkModel
+  type PasswordRecoveryModel,
+  type UseResendLinkModel,
+  type UserAuthModel,
+  type UserCreatePasswordModel,
+  type UserLoginModel,
+  type UserRegistrationModel,
 } from '../../types/auth'
 
 export const AuthService = {
-    me () {
-        return $api.get<UserAuthModel>('auth/me')
-    },
+  confirmEmail(confirmationCode: string) {
+    return $api.post('auth/registration-confirmation', { confirmationCode })
+  },
 
-    login (params: UserLoginModel) {
-        return $api.post<UserLoginModel, AxiosResponse<{ accessToken: string }>>('auth/login', params)
-            .then((response) => {
-                localStorage.setItem('token', response.data.accessToken)
-            })
-    },
+  createPassword(params: UserCreatePasswordModel) {
+    return $api.post('/auth/new-password', params)
+  },
 
-    logout () {
-        return $api.post('auth/logout').then(() => { localStorage.removeItem('token') })
-    },
+  login(params: UserLoginModel) {
+    return $api
+      .post<UserLoginModel, AxiosResponse<{ accessToken: string }>>('auth/login', params)
+      .then(response => {
+        localStorage.setItem('token', response.data.accessToken)
+      })
+  },
 
-    registration (params: UserRegistrationModel) {
-        return $api.post<null, AxiosResponse, UserRegistrationModel>('/auth/registration', params)
-    },
+  logout() {
+    return $api.post('auth/logout').then(() => {
+      localStorage.removeItem('token')
+    })
+  },
 
-    confirmEmail (confirmationCode: string) {
-        return $api.post('auth/registration-confirmation', { confirmationCode })
-    },
+  me() {
+    return $api.get<UserAuthModel>('auth/me')
+  },
 
-    createPassword (params: UserCreatePasswordModel) {
-        return $api.post('/auth/new-password', params).catch((e: AxiosError<{ message: string }>) => {
-            console.log(e.response?.data.message)
-        })
-    },
+  passwordRecovery(params: PasswordRecoveryModel) {
+    return $api.post('auth/password-recovery', params)
+  },
 
-    resendEmail (params: UseResendLinkModel) {
-        return $api.post('/auth/registration-email-resending', params)
-    },
+  registration(params: UserRegistrationModel) {
+    return $api.post<null, AxiosResponse, UserRegistrationModel>('/auth/registration', params)
+  },
 
-    passwordRecovery (params: PasswordRecoveryModel) {
-        return $api.post('auth/password-recovery', params)
-    }
+  resendEmail(params: UseResendLinkModel) {
+    return $api.post('/auth/registration-email-resending', params)
+  },
 }
